refactor: split services module
